3 курс 1 семестр/Код будущего/requirements.txt
annotated-types==0.6.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.104.1
h11==0.14.0
idna==3.4
pydantic==2.4.2
pydantic_core==2.11.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
starlette==0.31.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
